kernelci.legacy: move code to check for new commit
Move the code that checks whether a new commit is available based on a
file on the legacy kernelci-backend storage to the new kernelci.legacy
module.  This is in preparation for removal when the new API is ready
to be rolled out in production as it keeps track of the latest git
commit in the database rather than with a special file.
